Add the FaaS puzzle runtime.
This change introduces a parameterized runtime that abstracts across
multiple possible FaaS implementations, with a reference implementation
that uses AWS Lambda.

It also updates the evaluate model to pass through what is being
evaluated, be it the puzzle, the solution or a callback. This enables
the use of a single FaaS function for multiple use cases, which is good
for keeping the function hot, avoiding slowing down responses through
cold starts.