Run using an unprivileged database user
BREAKING CHANGE: Existing installations will need to configure DB
credentials per the upgrade instructions in the README

Modify the initial setup to include configuration of two database
passwords, one for the `postgres` user and one for a `hunter2` user.
For dev setups hardcode these strings to the imaginative `"hunter2"`
value.
For test we need to be able to create databases during the test run so
use the `postgres` user with a random password.
For production setups require passwords for both `postgres` and
`hunter2` users to be passed in via environment for the first launch.
Tolerate the `postgres` password being removed once the database has
been setup.