Don't use jQuery in Vue components/entrypoints
jQuery and Vue are known not to play well with each other. jQuery
expects to "own" the whole DOM and Vue expects to "own" the whole DOM
within its mountpoint.
Ultimately we probably want to move away from jQuery entirely. As a step
towards that, let's stop using it for AJAX calls inside of our Vue
components and use axios, which is recommended in the Vue docs, instead.