Improved DB migration instructions
Having actually migrated the DB for two prod sites I've come up with a
much better set of instructions for this. Sadly I'll never get to follow
them but maybe it will help someone else.
This also adds an extra user for the postgres exporter to use with
privilege to read all the things it needs delegated to it from the
superuser via stored procedure magic.