Use Progress model in Puzzle.answered_by
This changes the method to return a boolean instead of a list of
correct guesses - which was in all but two places simply tested
for truth.

In one other place this resulted in a test failure due to the test
altering a guess, which is not supported by the signal code.