Minimise database queries for site configuration
Add a middleware which stores the current site configuration on the
request object at the beginning of the request, then use this each time
we need to access the configuration, rather than requesting it again.
The one case where this doesn't work is where configuration is required
in the constructor of the `UserSignupForm`. This is a relatively
low-traffic view though, so it ought to be fine.