Fix hint timings in the team admin
When multiple guesses gave the team the unlock, the latest one was
being used as the basis for hint timings in the admin team page,
due to not correctly disambiguating the TeamUnlocks. This resulted in
hints which had been given to the team as being announced as being
due in the past.

Closes #428