Pin buildkit image to 0.9.3 and name build nodes
Pinning the buildkit image will prevent upstream updates breaking our
build. Right now buildx-stable-1 is v0.10.0 and it seems to have broken
us due to some obscure cache mount behaviour.
Naming the node with the version will ensure that bootstrapped build
containers are not re-used if they are not running the version we
desire.