Fix websocket tests under freezegun 1.1+
While freezegun is in use and time is frozen, time does not advance from
the point of view of the async event loop. Hence waiting for an event in
the future will hang forever. This commit removes all such waits.

This is somewhat fragile: anything which breaks these tests is liable to
cause them to hang, rather than fail. A future improvement could be to
allow time.monotonic to advance as real time does, but also to be ticked
forward - if that doesn't break anything.

The advantage of the new situation is that the tests are not dependent
on advancing real time at all so are less flaky.

Fixes #311