Fix Event script not appearing on many pages
This was due to those pages overriding the "script" block inherited from
the base template, which was done because these pages have their own
webpack bundles which should not be loaded alongside those from the base
template, which would load duplicates, which breaks things.

The fix was to create a block specifically for the webpack bundles,
which is then overridden without adding the super-block. The other block
can then be left un-overridden in these pages.