Require hints to be accepted
Store whether they have been accepted team-wide on TPP object. Per-user
would make some sense, but complicates things quite a bit and I think
it'd be weird to not have all team-members on the same page.

Changes to the admin views to reflect the additional info: Team Admin
page gets a CSS class to fade out unrevealed hints, whilst the Progress
page only displays its exclamation mark if there are no scheduled *and
no unrevealed* hints. The rationale here is that teams which are
actively engaged are probably going to accept their hints and admins
will want to let that happen before giving custom help.

Questionable: this is implemented with minimal change to the websocket
protocol: the hint object gains an "accepted" field, and hints are
initially sent without the hint text. Accepting a hint immediately
produces a second "new hint" message which on the client updates the hint
object with text.
This is as opposed to having separate "hint available" and "hint
accepted" message types. Maybe that's better for some reason.
One reason could be notifications: the person who accepted the hint
probably does not need a notification, but she will get one.

Closes: 510