Allow blank or non-unique clue text
Include a simple means of disambiguating clues in the admin, so that
they can be chosen from the list unambiguously.

Improve the layout of the clue list. This is partly due to supporting
blank unlocks, but goes beyond the minimum required to make that look
OK: the layout is now a grid, and restores the intuitive ordering of
guesses first, then the clues which they revealed.

Fixes highlighting new clues, which were broken in the ugprade to Vue3.

Fixes #378