Add hint obsoletion
Hints gain a new m2m field to Unlock. When any of the chosen unlocks is
unlocked by a team, the hint is immediately revealed, and styled to
indicate that the team no longer needs the associated information.

Some refactoring in consumers.py was necessary to preserve performance:

* schedule_hint_msg's message now takes a list of hint IDs so that
  multiple hints can be scheduled at once, which is significantly more
  efficient if a puzzle has several.
* schedule_hint first checks if the hint has a scheduled time to avoid
  doing more work if it doesn't
* schedule_hint takes the TPP object as a parameter so that it can be
  called multipled times with the same one.
* more prefetches

These are significant improvements with lots of hints, but it's less
clear how big they are with few hints. Still, we often have quite a few
hints on longer puzzles and it doesn't seem like this will have any
negative impact except for prefetches (which have little impact if the
prefetch is actually necessary and used, even if they don't cut any
queries out).

I think that this should probably be replaced soon (by tracking obsolete
state with another M2M field on TPP)