Fix progress page filtering highest-guessing team
This happened when the highest-guessing team guesses, increasing the
total and hence maximum for the total guesses filter. Code existed which
tried to increase the selected high end of the filter to the new maximum
if it was already on the old maximum. However, this was not working
because the value of the computed property max_total_guesses did not
update until after the updateData method returned.

Fix this by adding a pure function as a method on the component to
compute max total guesses and call this from the computed property, as
well as updateData.

Also do the same for the similar logic of the other filters.